+++ src/services/form-submissions.service.ts
@@ -1,6 +1,4 @@
-import { LandingConfig } from '@/providers/Questionnaire.provider';
 import { FormSubmission } from '@/types/forms.types';
-import { getLandingConfigKey } from '@/utils/local-storage';
 import {
   useMutation,
   useQueries,
@@ -9,11 +7,6 @@ import {
   UseQueryOptions,
 } from '@tanstack/react-query';
 import { fetchWithAuth } from './authorized-request.service';
-import {
-  ONBOARDING_SLUGS,
-  RECOMMENDATION_QUESTIONNAIRES,
-  VALUATION_QUESTIONNAIRE,
-} from './questionnaires.service';
 
 type GetFormSubmissionVariables = {
   formId: number;
@@ -68,50 +61,6 @@ export const useFormSubmissions = (variables: GetFormSubmissionVariables[]) => {
   return useQueries({
     queries: variables.map((v) => getFormSubmissionQueryVariables(v)),
   });
-};
-
-export const useOnboardingCompletion = () => {
-  const allOnboardingQuestionnaires = [
-    VALUATION_QUESTIONNAIRE,
-    ...RECOMMENDATION_QUESTIONNAIRES,
-  ];
-  const formSubmissionQueries = useFormSubmissions(
-    allOnboardingQuestionnaires
-      .map((q) => {
-        return q.forms.map((f) => {
-          return {
-            formId: f.id,
-          };
-        });
-      })
-      .flat()
-  );
-
-  // CALCULATED
-  // Only started onboarding when they have finished one of the recommendation forms, which excludes the valuation form (index 0)
-  const hasStartedOnboarding = formSubmissionQueries.some(
-    (query, index) => index !== 0 && !!query.data?.json_blob
-  );
-  const hasCompletedOnboarding = formSubmissionQueries.every((query) => {
-    return !!query.data?.json_blob;
-  });
-  const completion = allOnboardingQuestionnaires.map((q, index) => {
-    const hasFormSubmission = !!formSubmissionQueries[index].data?.json_blob;
-    const rawLandingConfig = localStorage.getItem(getLandingConfigKey(q));
-    const landingConfig = rawLandingConfig
-      ? (JSON.parse(rawLandingConfig) as LandingConfig)
-      : undefined;
-    const landingProgress =
-      ((landingConfig?.pageIndex ?? 0) / q.forms[0].definition.pages.length) *
-      100;
-    return {
-      title: q.forms[0].name,
-      completionPercentage: hasFormSubmission ? 100 : landingProgress,
-      href: `/q/${ONBOARDING_SLUGS[index]}`,
-    };
-  });
-
-  return { hasStartedOnboarding, hasCompletedOnboarding, completion };
 };
 
 type SubmitFormVariables = {
